+++ src/org/lwjgl/demo/opengl/shader/InfiniteDraggablePlaneDemo.java
@@ -32,7 +32,7 @@ public class InfiniteDraggablePlaneDemo {
     private long window;
     private int width = 1200;
     private int height = 800;
-    private int mouseX, mouseY;
+    private double mouseX, mouseY;
     private Vector3f dragStartWorldPos = new Vector3f();
     private Vector3f dragCamNormal = new Vector3f();
     private Vector3f dragRayOrigin = new Vector3f();
@@ -189,13 +189,13 @@ public class InfiniteDraggablePlaneDemo {
             drag(xpos, ypos);
         else if (viewing)
             rotate(xpos, ypos);
-        mouseX = (int) xpos;
-        mouseY = (int) ypos;
+        mouseX = xpos;
+        mouseY = ypos;
     }
 
     private void rotate(double xpos, double ypos) {
-        float deltaX = (float) xpos - mouseX;
-        float deltaY = (float) ypos - mouseY;
+        float deltaX = (float) (xpos - mouseX);
+        float deltaY = (float) (ypos - mouseY);
         vMat.rotateLocalY(deltaX * 0.005f);
         vMat.rotateLocalX(deltaY * 0.005f);
     }
@@ -235,7 +235,7 @@ public class InfiniteDraggablePlaneDemo {
     private void dragBegin() {
         dragging = true;
         vpMatDrag.set(vpMat);
-        vpMatDrag.unprojectRay(mouseX, height - mouseY, new int[] {0, 0, width, height}, dragRayOrigin, dragRayDir);
+        vpMatDrag.unprojectRay((float) mouseX, (float) (height - mouseY), new int[] {0, 0, width, height}, dragRayOrigin, dragRayDir);
         float t = intersectRayPlane(dragRayOrigin, dragRayDir, new Vector3f(), new Vector3f(0, dragRayOrigin.y > 0 ? 1 : -1, 0), 1E-5f);
         dragStartWorldPos.set(dragRayDir).mul(t).add(dragRayOrigin);
         vMat.positiveZ(dragCamNormal);
