+++ src/org/lwjgl/demo/vulkan/raytracing/SdfBricks.java
@@ -101,17 +101,17 @@ public class SdfBricks {
     private static final Vector3f tmpv3 = new Vector3f();
     private static final boolean[] keydown = new boolean[GLFW_KEY_LAST + 1];
     private static boolean mouseDown;
-    private static int mouseX, mouseY;
-  
+    private static double mouseX, mouseY;
+
     private static void onCursorPos(long window, double x, double y) {
         if (mouseDown) {
-            float deltaX = (float) x - mouseX;
-            float deltaY = (float) y - mouseY;
+            float deltaX = (float) (x - mouseX);
+            float deltaY = (float) (y - mouseY);
             viewMatrix.rotateLocalY(deltaX * 0.004f);
             viewMatrix.rotateLocalX(deltaY * 0.004f);
         }
-        mouseX = (int) x;
-        mouseY = (int) y;
+        mouseX = x;
+        mouseY = y;
     }
 
     private static void onKey(long window, int key, int scancode, int action, int mods) {
